Better layout for storybook
