+++ src/Pages/Authenticated/StorybookPage/Items/Greeting.tsx
@@ -4,8 +4,14 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
+import { CenteredRow } from '../Components'
+
 const Greeting = () => {
-  return <h3>Hello, choose component category!</h3>
+  return (
+    <CenteredRow>
+      <h1>Hello, choose component category!</h1>
+    </CenteredRow>
+  )
 }
 
 export default Greeting
