+++ src/Pages/Authenticated/StorybookPage/Items/IndicatorsBook.tsx
@@ -7,18 +7,19 @@
 import { LayoutRow } from '../../Components/Layout/Layout'
 import { CircleIndicator } from '../../SettingsPage/Tabs/Advanced/CircleIndicator'
 import { Variant } from '../../SettingsPage/Tabs/Advanced/types'
+import { CenteredRow } from '../Components'
 
 const IndicatorsBook = () => {
   return (
-    <>
-      {[12, 16, 20, 24].map((s) => (
+    <CenteredRow>
+      {[12, 16, 20, 24, 28, 34, 38, 42].map((s) => (
         <LayoutRow key={`cid-${s}-s`}>
           {['ok', 'warning', 'error'].map((v) => (
             <CircleIndicator key={`${v}-${s}`} variant={v as Variant} size={s} />
           ))}
         </LayoutRow>
       ))}
-    </>
+    </CenteredRow>
   )
 }
