+++ src/Pages/Authenticated/StorybookPage/StorybookPage.tsx
@@ -9,44 +9,7 @@ import React, { useMemo } from 'react'
 import { NavLinkTabs } from '../../../Components/Tabs/NavLinkTabs'
 import ROUTES from '../../../constants/routes'
 import { useLocation } from 'react-router-dom'
-
-interface RouteInfo {
-  title: string
-  component: string
-}
-
-const subRoute = (subRoute: string): string => `${ROUTES.STORYBOOK}${subRoute}`
-
-export const STORYBOOK_ROUTES: { [key: string]: RouteInfo } = {
-  [subRoute('/buttons')]: {
-    component: './Items/ButtonsBook',
-    title: 'Buttons',
-  },
-  [subRoute('/indicators')]: {
-    component: './Items/IndicatorsBook',
-    title: 'Indicators',
-  },
-  [subRoute('/nat-tooltips')]: {
-    component: './Items/TooltipsBook',
-    title: 'Tooltips',
-  },
-  [subRoute('/notifications')]: {
-    component: './Items/NotificationsBook',
-    title: 'Notifications',
-  },
-  [subRoute('/inputs')]: {
-    component: './Items/InputsBook',
-    title: 'Inputs',
-  },
-  [subRoute('/password-change')]: {
-    component: './Items/PasswordChangeBook',
-    title: 'PasswordChange',
-  },
-  [subRoute('/progress-bar')]: {
-    component: './Items/ProgressBarBook',
-    title: 'Progress',
-  },
-}
+import { STORYBOOK_ROUTES } from './storybook.utils'
 
 const toTabs = () =>
   Object.keys(STORYBOOK_ROUTES).map((path) => ({
